refactor(fasthttphandler): Remove unnecessary refactors
