+++ fasthttpadaptor/adaptor.go
@@ -5,7 +5,9 @@ package fasthttpadaptor
 import (
 	"bufio"
 	"io"
+	"net"
 	"net/http"
+	"sync"
 
 	"github.com/valyala/fasthttp"
 )
@@ -62,9 +64,9 @@ func NewFastHTTPHandler(h http.Handler) fasthttp.RequestHandler {
 		go func() {
 			h.ServeHTTP(w, r.WithContext(ctx))
 			w.chOnce.Do(func() {
-				w.modeCh <- modeCopy
+				w.modeCh <- modeDone
 			})
-			w.close()
+			w.Close()
 
 			// Wait for the net/http handler to complete before releasing.
 			// (e.g. wait for hijacked connection)
@@ -73,7 +75,7 @@ func NewFastHTTPHandler(h http.Handler) fasthttp.RequestHandler {
 		}()
 
 		switch <-w.modeCh {
-		case modeCopy:
+		case modeDone:
 			// No flush occurred before the handler returned.
 			// Send the data as one chunk.
 			ctx.SetStatusCode(w.StatusCode())
@@ -107,7 +109,7 @@ func NewFastHTTPHandler(h http.Handler) fasthttp.RequestHandler {
 			// Signal that the net/http -> fasthttp copy is complete.
 			w.wg.Done()
 
-		case modeStream:
+		case modeFlushed:
 			// Flush occurred before handler returned.
 			// Send the first 512 bytes and start streaming
 			// the rest of the first chunk and new data as it arrives.
@@ -202,7 +204,7 @@ func NewFastHTTPHandler(h http.Handler) fasthttp.RequestHandler {
 			w.streamCond.Signal()
 			w.streamCond.L.Unlock()
 
-		case modeHijack:
+		case modeHijacked:
 			// The net/http handler called w.Hijack().
 			// Copy data bidirectionally between the
 			// net/http and fasthttp connections.
@@ -230,4 +232,223 @@ func NewFastHTTPHandler(h http.Handler) fasthttp.RequestHandler {
 			w.wg.Done()
 		}
 	}
+}
+
+// Use a minimum buffer size of 32 KiB.
+const minBufferSize = 32 * 1024
+
+var bufferPool = &sync.Pool{
+	New: func() any {
+		b := make([]byte, minBufferSize)
+		return &b
+	},
+}
+
+var writerPool = &sync.Pool{
+	New: func() any {
+		pr, pw := io.Pipe()
+		return &responseWriter{
+			h:            make(http.Header),
+			r:            pr,
+			w:            pw,
+			modeCh:       make(chan modeType),
+			responseBody: acquireBuffer(),
+			streamCond:   sync.NewCond(&sync.Mutex{}),
+		}
+	},
+}
+
+// modeType are all possible responseWriter modes.
+type modeType int
+
+const (
+	modeUnknown modeType = iota //nolint:unused
+
+	// modeDone means a response should directly copy the
+	// net/http data into fashttp.RequestCtx on completion.
+	modeDone
+
+	// modeFlushed means a response should send a buffered stream
+	// of data every time responseWriter.Flush() is called.
+	modeFlushed
+
+	// modeHijacked means that a response should do nothing else but
+	// copy data directly into the client's raw connection.
+	modeHijacked
+)
+
+// responseWriter represents a net/http adaptor that implements
+// the http.ResponseWriter, http.Flusher, and http.Hijacker interfaces.
+type responseWriter struct {
+	handlerConn  net.Conn
+	ctx          *fasthttp.RequestCtx
+	h            http.Header
+	r            *io.PipeReader
+	w            *io.PipeWriter
+	modeCh       chan modeType
+	responseBody *[]byte
+	streamCond   *sync.Cond
+	statusCode   int
+	chOnce       sync.Once
+	closeOnce    sync.Once
+	isStreaming  bool
+	wg           sync.WaitGroup
+	hijackedWg   sync.WaitGroup
+}
+
+// acquireResponseWriter returns a pointer to a slice of 0 length and
+// at least minBufferSize capacity.
+func acquireResponseWriter(ctx *fasthttp.RequestCtx) *responseWriter {
+	w, ok := writerPool.Get().(*responseWriter)
+	if !ok {
+		panic("fasthttpadaptor: cannot get *responseWriter from writerPool")
+	}
+	w.reset()
+
+	w.ctx = ctx
+	return w
+}
+
+// releaseResponseWriter recycles the buffer for reuse.
+func releaseResponseWriter(w *responseWriter) {
+	releaseBuffer(w.responseBody)
+	writerPool.Put(w)
+}
+
+// acquireBuffer returns a pointer to a slice of 0 length and
+// at least minBufferSize capacity.
+func acquireBuffer() *[]byte {
+	buf, ok := bufferPool.Get().(*[]byte)
+	if !ok {
+		panic("fasthttpadaptor: cannot get *[]byte from bufferPool")
+	}
+
+	*buf = (*buf)[:0]
+	return buf
+}
+
+// releaseBuffer recycles the buffer for reuse.
+func releaseBuffer(buf *[]byte) {
+	bufferPool.Put(buf)
+}
+
+// StatusCode returns the response's status code.
+//
+// If no status code is set, it returns http.StatusOK by default.
+func (w *responseWriter) StatusCode() int {
+	if w.statusCode == 0 {
+		return http.StatusOK
+	}
+	return w.statusCode
+}
+
+// Header returns the current response header.
+func (w *responseWriter) Header() http.Header {
+	return w.h
+}
+
+// WriteHeader sets the response's status code.
+func (w *responseWriter) WriteHeader(statusCode int) {
+	w.statusCode = statusCode
+}
+
+// Write writes the data to the connection.
+//
+// Write supports both buffered mode and streaming mode.
+func (w *responseWriter) Write(p []byte) (int, error) {
+	w.streamCond.L.Lock()
+	defer w.streamCond.L.Unlock()
+
+	if w.isStreaming {
+		// Streaming mode is on.
+		// Stream directly to the conn writer.
+		return w.w.Write(p)
+	}
+
+	// Streaming mode is off.
+	// Write to the response body buffer for flushing later.
+	*w.responseBody = append(*w.responseBody, p...)
+	return len(p), nil
+}
+
+// Flush signals streaming mode.
+//
+// To ensure proper setup when using responseWriter.Write,
+// Flush blocks until streaming mode is ready.
+func (w *responseWriter) Flush() {
+	// Trigger streaming mode setup.
+	w.chOnce.Do(func() {
+		w.modeCh <- modeFlushed
+	})
+
+	// Wait for streaming mode.
+	w.streamCond.L.Lock()
+	defer w.streamCond.L.Unlock()
+	for !w.isStreaming {
+		w.streamCond.Wait()
+	}
+}
+
+// Hijack signals hijack mode.
+//
+// When called, the net/http handler assumes full control over the returned connection.
+func (w *responseWriter) Hijack() (net.Conn, *bufio.ReadWriter, error) {
+	// Prevent fasthttp from closing it or
+	// doing anything else with it.
+	w.ctx.HijackSetNoResponse(true)
+
+	netHTTPConn, fasthttpConn := net.Pipe()
+	w.handlerConn = fasthttpConn
+
+	// Trigger hijacked mode.
+	w.chOnce.Do(func() {
+		w.modeCh <- modeHijacked
+	})
+
+	bufRW := bufio.NewReadWriter(bufio.NewReader(netHTTPConn), bufio.NewWriter(netHTTPConn))
+
+	// Write any unflushed body to the hijacked connection buffer.
+	if len(*w.responseBody) > 0 {
+		_, _ = bufRW.Write(*w.responseBody)
+		_ = bufRW.Flush()
+	}
+	return netHTTPConn, bufRW, nil
+}
+
+// Close signals that the net/http handler
+// finished using the writer.
+//
+// When called in streaming mode, the responseWriter can flush the remaining unread data
+// and Close the client connection. Otherwise, this method does nothing.
+func (w *responseWriter) Close() error {
+	w.closeOnce.Do(func() {
+		if w.w != nil {
+			_ = w.w.Close()
+		}
+	})
+	return nil
+}
+
+// reset clears data from a recycled responseWriter.
+func (w *responseWriter) reset() {
+	// Note: reset() must only run after a fasthttp handler finishes
+	// proxying the full net/http handler response to ensure no data races.
+	w.ctx = nil
+	w.statusCode = 0
+
+	w.w = nil
+	w.r = nil
+	w.handlerConn = nil
+
+	// Clear the http Header
+	for key := range w.h {
+		delete(w.h, key)
+	}
+
+	// Get a new buffer for the response body
+	w.responseBody = acquireBuffer()
+
+	w.chOnce = sync.Once{}
+	w.closeOnce = sync.Once{}
+	w.isStreaming = false
 }
